trying to process era5 data

+++ tcpips/era5.py
@@ -11,7 +11,8 @@ import cdsapi
 from typing import List
 import xarray as xr
 from sithom.time import timeit
-from .constants import ERA5_PATH
+from .constants import ERA5_RAW_PATH, ERA5_PI_OG_PATH, ERA5_PI_PATH
+from .pi import calculate_pi
 
 
 @timeit
@@ -30,7 +31,7 @@ def download_single_levels(
         output_file (str): Name of the output file to save the data.
         redo (bool): If True, download the data again even if the file already exists. Default is False.
     """
-    if os.path.exists(os.path.join(ERA5_PATH, output_file)) and not redo:
+    if os.path.exists(os.path.join(ERA5_RAW_PATH, output_file)) and not redo:
         print(f"File {output_file} already exists. Use redo=True to download again.")
         return
     c = cdsapi.Client()
@@ -48,15 +49,40 @@ def download_single_levels(
             "month": months,
             "time": "00:00",
         },
-        os.path.join(ERA5_PATH, output_file),
+        os.path.join(ERA5_RAW_PATH, output_file),
     )
     print(f"Downloaded single-level data to {output_file}")
     return
 
 
+def break_single_levels(years: List[str], output_file: str) -> None:
+    """
+    Breaks the single-level data into decade chunks and saves them as separate files.
+
+    Args:
+        years (List[str]): List of years to download data for.
+        output_file (str): Name of the output file to save the data.
+    """
+    # break the single level file into decade chunks
+    # and save them as separate files
+    prefix = output_file.replace(".nc", "")
+    ds = xr.open_dataset(os.path.join(ERA5_RAW_PATH, output_file))
+    for i in range(0, len(years), 10):
+        j = min(i + 10, len(years))
+        file_name = f"{prefix}_years{years[i]}_{years[j-1]}.nc"
+        file_path = os.path.join(ERA5_RAW_PATH, file_name)
+        out_ds = ds.sel(valid_time=slice(f"{years[i]}-01-01", f"{years[j-1]}-12-31"))
+        out_ds.to_netcdf(file_path)
+        print(f"Saved {file_path}")
+
+
 @timeit
 def download_pressure_levels(
-    years: List[str], months: List[str], output_file: str, redo: bool = False
+    years: List[str],
+    months: List[str],
+    output_file: str,
+    redo: bool = False,
+    stitch_together: bool = True,
 ) -> None:
     """
     Downloads monthly-averaged pressure-level fields:
@@ -76,8 +102,9 @@ def download_pressure_levels(
         months (List[str]): List of months to download data for.
         output_file (str): Name of the output file to save the data.
         redo (bool): If True, download the data again even if the file already exists. Default is False.
+        stitch_together (bool): If True, stitch the temporary files together at the end. Default is True.
     """
-    if os.path.exists(os.path.join(ERA5_PATH, output_file)) and not redo:
+    if os.path.exists(os.path.join(ERA5_RAW_PATH, output_file)) and not redo:
         print(f"File {output_file} already exists. Use redo=True to download again.")
         return
     print(
@@ -93,7 +120,7 @@ def download_pressure_levels(
             file_name = (
                 f"{output_file.replace('.nc', '')}_years{years[i]}_{years[j-1]}.nc"
             )
-            file_path = os.path.join(ERA5_PATH, file_name)
+            file_path = os.path.join(ERA5_RAW_PATH, file_name)
             file_paths.append(file_path)
 
             # check if the temporary sfile already exists (this would indicate the script crashed partway through last time)
@@ -102,22 +129,23 @@ def download_pressure_levels(
             else:
                 download_pressure_levels(years[i:j], months, file_name)
         # stitch the files together
-        output_path = os.path.join(ERA5_PATH, output_file)
+        output_path = os.path.join(ERA5_RAW_PATH, output_file)
         # append the files together along the time dimension and save to output_file
         # using xarray
-        print(f"Stitching files together: {file_paths} to {output_path}")
-        ds = xr.open_mfdataset(
-            file_paths,
-            combine="nested",
-            concat_dim="time",
-            parallel=True,
-        )
-        ds.to_netcdf(output_path)
-        # remove the temporary files
-        if os.path.exists(output_path):
-            for file_path in file_paths:
-                os.remove(file_path)
-        print(f"Downloaded pressure-level data to {output_path}")
+        if stitch_together:
+            print(f"Stitching files together: {file_paths} to {output_path}")
+            ds = xr.open_mfdataset(
+                file_paths,
+                combine="nested",
+                concat_dim="valid_time",
+                parallel=True,
+            )
+            ds.to_netcdf(output_path)
+            # remove the temporary files
+            if os.path.exists(output_path):
+                for file_path in file_paths:
+                    os.remove(file_path)
+            print(f"Downloaded pressure-level data to {output_path}")
         return
     # if len(years) <= 10, download the data normally using cdsapi
 
@@ -151,7 +179,7 @@ def download_pressure_levels(
             "month": months,
             "time": "00:00",
         },
-        os.path.join(ERA5_PATH, output_file),
+        os.path.join(ERA5_RAW_PATH, output_file),
     )
     print(f"Downloaded pressure-level data to {output_file}")
     return
@@ -160,7 +188,8 @@ def download_pressure_levels(
 def download_era5_data() -> None:
     """
     Downloads ERA5 data for the specified years and months.
-    This function is a wrapper around the download_single_levels and download_pressure_levels functions.
+    This function is a wrapper around the download_single_levels
+    and download_pressure_levels functions.
     """
     # Specify the desired years and months.
     years = [
@@ -171,13 +200,56 @@ def download_era5_data() -> None:
     # Download the single-level (surface) variables.
     download_single_levels(years, months, "era5_single_levels.nc", redo=False)
 
+    # break the single level file into decade chunks
+    # and save them as separate files
+    break_single_levels(years, "era5_single_levels.nc")
+
     # Download the pressure-level variables (including geopotential).
-    download_pressure_levels(years, months, "era5_pressure_levels.nc", redo=True)
+    download_pressure_levels(
+        years, months, "era5_pressure_levels.nc", redo=False, stitch_together=False
+    )
+
+
+def era5_pi_decade(single_level_path: str, pressure_level_path: str) -> None:
+    years_str = single_level_path.split("years")[1].replace(".nc", "")
+    single_ds = xr.open_dataset(single_level_path)
+    pressure_ds = xr.open_dataset(pressure_level_path)
+    # let's assume they are on the same grid
+    # and that the time dimension is the same
+    combined_ds = xr.merge([single_ds, pressure_ds])
+    # convert units and rename variables here (not implemented)
+    # combined_ds = convert_units(combined_ds)
+    # calculate potential intensity
+    # pi_ds = calculate_pi(combined_ds)
+    # save the potential intensity dataset
+    # pi_ds.to_netcdf(os.path.join(ERA5_PI_OG_PATH, "era5_pi.nc"))
+    combined_ds.to_netcdf(os.path.join(ERA5_PI_OG_PATH, f"era5_pi_years{years_str}.nc"))
+
+
+def era5_pi(years: List[str]) -> None:
+    """
+    Calculate potential intensity (PI) using the downloaded ERA5 data.
+    This function is a placeholder and should be implemented with the actual calculation logic.
+    """
+    # Placeholder for potential intensity calculation logic
+    # First convert the data
+    for i in range(0, len(years), 10):
+        j = min(i + 10, len(years))
+        single_level_path = os.path.join(
+            ERA5_RAW_PATH, f"era5_single_levels_years{years[i]}_{years[j-1]}.nc"
+        )
+        pressure_level_path = os.path.join(
+            ERA5_RAW_PATH, f"era5_pressure_levels_years{years[i]}_{years[j-1]}.nc"
+        )
+        era5_pi_decade(single_level_path, pressure_level_path)
 
 
 if __name__ == "__main__":
     # python -m tcpips.era5
     download_era5_data()
+    era5_pi(
+        [str(year) for year in range(1980, 2025)]
+    )  # Modify or extend this list as needed.)
     # problem: the era5 pressure level data is too big to save in one file
     # so I have split it into chunks of 10 years.
     # This means that future scripts also need to be able to handle this.
